Fix doc update bug
