+++ functions/src/aqi-calculation/util.ts
@@ -54,8 +54,9 @@ enum InvalidAqiErrors {
  * @returns an array for each `InvalidAqiErrors` with each error set to `false`.
  */
 function getDefaultInvalidAqiErrors(): boolean[] {
-  // TypeScript does not provide a way to get the number of elements in an enum,
-  // so this gets the number of elements using the enum reverse mapping.
+  // TypeScript does not provide a way to get the number of elements in an
+  // enumeration, so this gets the number of elements using the enumeration
+  // reverse mapping.
   // eslint-disable-next-line no-magic-numbers
   const invalidAqiErrorCount = Object.keys(InvalidAqiErrors).length / 2;
   return new Array<boolean>(invalidAqiErrorCount).fill(false);
